requirements.txt
discord.py
discord-components
autopep8
deta
python-dotenv
